Discord.Net: Fixed data scheme of user objects
Formerly user objects store the discriminator value as an unsigned short number, which I think is possibly for memory optimization. To detect the single-zero discriminator and to allow the absence of the discriminator field in the future API, we make it store an optional string instead.

The current discriminator property and the ushort one are kept for backward compatibility. The string discriminator property returns an empty string and the unsigned short property returns 10000 if the discriminator field in the original model is absent.